pipeline/requirements.txt
cycler==0.10.0
et-xmlfile==1.0.1
jdcal==1.3
matplotlib==2.0.0
numpy
openpyxl==2.4.8
pandas
pymongo
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
urllib3==1.25.3
biopython==1.70
